Feats(CRUD) : update,get ALL,get by Id, DELETE by id

+++ dvdstore_backend2/src/main/java/com/dvdstore/panier/controllers/PanierDto.java
@@ -2,5 +2,5 @@ package com.dvdstore.panier.controllers;
 
 import java.util.Date;
 
-public record PanierDto(Long idClient,Long idDvd,float prixUnitaire,float total,Date date,Long quantity) {
+public record PanierDto(Integer idClient, Integer idDvd, float prixUnitaire, float total, Date date, Integer quantity) {
 }
